Added UI Client

+++ Rail.DigitalTwin.Core/Models/DeviceModel.cs
@@ -3,7 +3,7 @@ namespace Rail.DigitalTwin.Core.Models
 {
     public class DeviceModel
     {
-        public DeviceModel(string deviceID, int timeElapsed, double latitude, double longitude) 
+        public DeviceModel(string deviceID, double timeElapsed, double latitude, double longitude) 
         { 
             DeviceID = deviceID;
             TimeElapsed = timeElapsed;
@@ -12,7 +12,7 @@ namespace Rail.DigitalTwin.Core.Models
         }
 
         public string DeviceID { get; set; }
-        public int TimeElapsed { get; set; }
+        public double TimeElapsed { get; set; }
         public double Latitude { get; set; }
         public double Longitude { get; set; }
     }
